[TM4C123_ACOMP] Updated EEPROM FLASH HIB and SysTick

+++ TM4C123_ACOMP/xDriver_MCU/Core/SYSTICK/Driver/Intrinsics/xSource/SYSTICK_TimeUsBase.c
@@ -20,5 +20,5 @@ void SysTick__vSetTimeUsBase(float32_t fBase)
 
 float32_t SysTick__fGetTimeUsBase(void)
 {
-    return (float32_t) SysTick_fTimeUsBase;
+    return SysTick_fTimeUsBase;
 }
